Added command 'version'

+++ src/commands/biwak/candump.cpp
@@ -49,7 +49,7 @@ int CCommandCanDump::exec(int argc, const char *argv[]) const /* virtual */
          
          while( ( message=m_can.m_rxBuffer.frontEntry() ) )
          {
-            printf( "  can0  %3X   [%d]  ", (int)message->getId(), message->getLen() );
+            printf( "  can0  %3X   [%d]  ", message->getId(), message->getLen() );
             for(int i1=0; i1<message->getLen(); i1++)
             {
                printf( "%02X ", message->getData()[i1]);
